restrict mirrors of any union type
